mesa-hostmot2 driver:  Remove int->double conversions
Fixes 'Unknown symbol __fixunsdfdi' on 32-bit gcc-4.[34].x kthreads

Also see:

http://permalink.gmane.org/gmane.linux.distributions.emc.user/18718
http://permalink.gmane.org/gmane.linux.distributions.emc.devel/3663

(Several more in components/lcd.c parse_fmt() that'll be harder/uglier
to squash)

+++ src/hal/drivers/mesa-hostmot2/watchdog.c
@@ -309,7 +309,8 @@ void hm2_watchdog_force_write(hostmot2_t *hm2) {
         // watchdog is disabled, MSb=1 is secret handshake with FPGA
         hm2->watchdog.timer_reg[0] = 0x80000000;
     } else {
-        tmp = (hm2->watchdog.instance[0].hal.param.timeout_ns * ((double)hm2->watchdog.clock_frequency / (double)(1000 * 1000 * 1000))) - 1;
+        tmp = (hm2->watchdog.instance[0].hal.param.timeout_ns *
+              hm2->watchdog.clock_frequency / (1000 * 1000 * 1000)) - 1;
         if (tmp < 0x80000000) {
             hm2->watchdog.timer_reg[0] = tmp;
         } else {
